requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
aiofiles==23.2.1
loguru==0.7.2
urllib3==2.1.0
certifi==2023.11.17
async-timeout==4.0.3
requests==2.31.0
selenium==4.11.0
webdriver-manager==3.8.6
schedule==1.2.1